+++ typescript/sdk/src/middleware/deploy.ts
@@ -1,15 +1,13 @@
 import { ethers } from 'ethers';
 
 import { ProxyAdmin__factory } from '@hyperlane-xyz/core';
-
 import {
-  InterchainAccountContracts,
-  InterchainAccountFactories,
-  InterchainQueryContracts,
-  InterchainQueryFactories,
-  interchainAccountFactories,
-  interchainQueryFactories,
-} from '../middleware';
+  InterchainAccountRouter,
+  InterchainAccountRouter__factory,
+  InterchainQueryRouter,
+  InterchainQueryRouter__factory,
+} from '@hyperlane-xyz/core';
+
 import { MultiProvider } from '../providers';
 import {
   HyperlaneRouterDeployer,
@@ -18,6 +16,25 @@ import {
   RouterFactories,
 } from '../router';
 import { ChainMap, ChainName } from '../types';
+
+export type InterchainAccountFactories =
+  RouterFactories<InterchainAccountRouter>;
+
+export const interchainAccountFactories: InterchainAccountFactories = {
+  router: new InterchainAccountRouter__factory(),
+};
+
+export type InterchainAccountContracts =
+  ProxiedRouterContracts<InterchainAccountRouter>;
+
+export type InterchainQueryFactories = RouterFactories<InterchainQueryRouter>;
+
+export const interchainQueryFactories: InterchainQueryFactories = {
+  router: new InterchainQueryRouter__factory(),
+};
+
+export type InterchainQueryContracts =
+  ProxiedRouterContracts<InterchainQueryRouter>;
 
 export abstract class MiddlewareRouterDeployer<
   MiddlewareRouterConfig extends RouterConfig,
