+++ typescript/sdk/src/middleware/liquidity-layer/LiquidityLayerApp.ts
@@ -11,11 +11,11 @@ import { utils } from '@hyperlane-xyz/utils';
 
 import { HyperlaneApp } from '../../HyperlaneApp';
 import { Chains } from '../../consts';
-import { LiquidityLayerContracts } from '../../middleware';
 import { MultiProvider } from '../../providers';
 import { ChainMap, ChainName } from '../../types';
 
 import { BridgeAdapterConfig } from './LiquidityLayerRouterDeployer';
+import { LiquidityLayerContracts } from './contracts';
 
 const PORTAL_VAA_SERVICE_TESTNET_BASE_URL =
   'https://wormhole-v2-testnet-api.certus.one/v1/signed_vaa/';
